requirements.txt
aiocache==0.12.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
black==23.11.0
botocore==1.33.5
bson==0.5.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
ghp-import==2.1.0
gql==3.4.1
graphql-core==3.2.3
griffe==0.38.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.5.1
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocs-material==9.4.14
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.0
mkdocstrings-python==1.7.5
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
paginate==0.5.6
pathspec==0.11.2
platformdirs==4.0.0
pydantic_core==2.14.5
Pygments==2.17.2
pymdown-extensions==10.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
typing_extensions==4.8.0
urllib3==1.26.18
watchdog==3.0.0
websockets==10.4
yarl==1.9.3
pymediasoup
